noticed something to refactor while writing the report
