+++ android/src/com/mygdx/game/AndroidLauncher.java
@@ -4,7 +4,6 @@ import android.os.Bundle;
 
 import com.badlogic.gdx.backends.android.AndroidApplication;
 import com.badlogic.gdx.backends.android.AndroidApplicationConfiguration;
-import com.mygdx.game.Application;
 
 public class AndroidLauncher extends AndroidApplication {
 	@Override
