+++ core/src/com/mygdx/game/ECS/EntityUtils/templates/Projectiles/AbstractProjectile.java
@@ -3,12 +3,12 @@ package com.mygdx.game.ECS.EntityUtils.templates.Projectiles;
 
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.math.Vector2;
-import com.mygdx.game.ECS.components.Box2DComponent;
-import com.mygdx.game.ECS.components.PositionComponent;
+import com.mygdx.game.ECS.components.misc.Box2DComponent;
+import com.mygdx.game.ECS.components.misc.PositionComponent;
 import com.mygdx.game.ECS.EntityUtils.strategies.Projectile.ProjectileType;
 import com.mygdx.game.ECS.components.projectiles.ProjectileComponent;
 import com.mygdx.game.ECS.components.flags.RenderComponent;
-import com.mygdx.game.ECS.components.SpriteComponent;
+import com.mygdx.game.ECS.components.misc.SpriteComponent;
 import com.mygdx.game.ECS.EntityUtils.templates.AbstractEntity;
 
 import static com.mygdx.game.utils.B2DConstants.BIT_GROUND;
