+++ core/src/com/mygdx/game/ECS/managers/ECSManager.java
@@ -10,19 +10,19 @@ import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.physics.box2d.Body;
 import com.badlogic.gdx.physics.box2d.Fixture;
 import com.mygdx.game.ECS.EntityUtils.EntityTemplateMapper;
-import com.mygdx.game.ECS.components.CollisionComponent;
+import com.mygdx.game.ECS.components.misc.CollisionComponent;
 import com.mygdx.game.ECS.components.flags.MovementControlComponent;
-import com.mygdx.game.ECS.components.ParentComponent;
+import com.mygdx.game.ECS.components.misc.ParentComponent;
 import com.mygdx.game.ECS.components.flags.PowerUpComponent;
 import com.mygdx.game.ECS.components.projectiles.ProjectileComponent;
-import com.mygdx.game.ECS.components.ShootingComponent;
-import com.mygdx.game.ECS.components.Box2DComponent;
-import com.mygdx.game.ECS.components.FontComponent;
-import com.mygdx.game.ECS.components.HealthComponent;
+import com.mygdx.game.ECS.components.misc.ShootingComponent;
+import com.mygdx.game.ECS.components.misc.Box2DComponent;
+import com.mygdx.game.ECS.components.misc.FontComponent;
+import com.mygdx.game.ECS.components.misc.HealthComponent;
 import com.mygdx.game.ECS.components.flags.PlayerComponent;
-import com.mygdx.game.ECS.components.PositionComponent;
-import com.mygdx.game.ECS.components.SpriteComponent;
-import com.mygdx.game.ECS.components.VelocityComponent;
+import com.mygdx.game.ECS.components.misc.PositionComponent;
+import com.mygdx.game.ECS.components.misc.SpriteComponent;
+import com.mygdx.game.ECS.components.misc.VelocityComponent;
 import com.mygdx.game.ECS.systems.AimingSystem;
 import com.mygdx.game.ECS.systems.MovementSystem;
 import com.mygdx.game.ECS.systems.ShootingSystem;
