+++ ios/src/com/mygdx/game/IOSLauncher.java
@@ -5,7 +5,6 @@ import org.robovm.apple.uikit.UIApplication;
 
 import com.badlogic.gdx.backends.iosrobovm.IOSApplication;
 import com.badlogic.gdx.backends.iosrobovm.IOSApplicationConfiguration;
-import com.mygdx.game.Application;
 
 public class IOSLauncher extends IOSApplication.Delegate {
     @Override
